Decouple code generation stage from M4 (#1002)
* Decouple code generation stage from M4

Removed code generator's dependency on M4, replacing it with an input
agnostic abstraction layer (gocodemodel.ts).
The naming and transformation phase still interface with M4.
At the end of transformation, the transformed M4 code model is converted
into a GoCodeModel which is then sent to the code generator.
Updated applicable references from gotest and in one case copied the old
code over which still talks to M4.

* propagate module and moduleVersion from session info

ensure that module and containing-module are mutually exclusive

* fix lint

* get azure-arm setting from session

* improve delimited parameters

* pass when there are no unit tests, clean up more lint

* remove duplicated code

* move comment to ctor

* minor cleanup

+++ packages/autorest.go/test/autorest/numbergroup/zz_number_client.go
@@ -761,7 +761,7 @@ func (client *NumberClient) putBigDecimalNegativeDecimalCreateRequest(ctx contex
 		return nil, err
 	}
 	req.Raw().Header["Accept"] = []string{"application/json"}
-	if err := runtime.MarshalAsJSON(req, -99999999.99); err != nil {
+	if err := runtime.MarshalAsJSON(req, float64(-99999999.99)); err != nil {
 		return nil, err
 	}
 	return req, nil
@@ -802,7 +802,7 @@ func (client *NumberClient) putBigDecimalPositiveDecimalCreateRequest(ctx contex
 		return nil, err
 	}
 	req.Raw().Header["Accept"] = []string{"application/json"}
-	if err := runtime.MarshalAsJSON(req, 99999999.99); err != nil {
+	if err := runtime.MarshalAsJSON(req, float64(99999999.99)); err != nil {
 		return nil, err
 	}
 	return req, nil
@@ -884,7 +884,7 @@ func (client *NumberClient) putBigDoubleNegativeDecimalCreateRequest(ctx context
 		return nil, err
 	}
 	req.Raw().Header["Accept"] = []string{"application/json"}
-	if err := runtime.MarshalAsJSON(req, -99999999.99); err != nil {
+	if err := runtime.MarshalAsJSON(req, float64(-99999999.99)); err != nil {
 		return nil, err
 	}
 	return req, nil
@@ -925,7 +925,7 @@ func (client *NumberClient) putBigDoublePositiveDecimalCreateRequest(ctx context
 		return nil, err
 	}
 	req.Raw().Header["Accept"] = []string{"application/json"}
-	if err := runtime.MarshalAsJSON(req, 99999999.99); err != nil {
+	if err := runtime.MarshalAsJSON(req, float64(99999999.99)); err != nil {
 		return nil, err
 	}
 	return req, nil
